fix: address remaining PR #29 review issues
- Change m_pendingCount from quint64 to int in all three services to
  make underflow visible to ASan/UBSan; add Q_ASSERT before each
  decrement to catch stale signal deliveries immediately
- Narrow Phase 5 genre comment in BookDiscoveryService to describe only
  the loop-body semantics, not the behaviour of a failed db.commit()
- Add QueryWorker::conn() helper and replace 16 QLatin1String(kConnectionName)
  call sites so any rename touches one place

+++ src/gui/services/search_service.h
@@ -79,7 +79,7 @@ signals:
 
 private:
     std::atomic<quint64> m_nextRequestId{1};
-    quint64              m_pendingCount{0};
+    int                  m_pendingCount{0};
 };
 
 // ---------------------------------------------------------------------------
